+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
-- Initial version.
+- Dart & Flutter 3.
+- Supports SOCKS version 5 protocol.
+- SSL support.
+- Supports ElectrumX and Fulcrum servers via socket(s).
+- Async support for non-blocking network communication.
+- Lightweight and minimal dependencies.
+- Working example.